move VA block updating to MutliblockWithDisplayBase super class
